+++ packages/query-core/src/__tests__/streamedQuery.test.tsx
@@ -128,6 +128,35 @@ describe('streamedQuery', () => {
     unsubscribe()
   })
 
+  test('should handle empty streams', async () => {
+    const key = queryKey()
+
+    const observer = new QueryObserver(queryClient, {
+      queryKey: key,
+      queryFn: streamedQuery({
+        streamFn: async function* () {},
+      }),
+    })
+
+    const unsubscribe = observer.subscribe(vi.fn())
+
+    expect(observer.getCurrentResult()).toMatchObject({
+      status: 'pending',
+      fetchStatus: 'fetching',
+      data: undefined,
+    })
+
+    await vi.advanceTimersByTimeAsync(50)
+
+    expect(observer.getCurrentResult()).toMatchObject({
+      status: 'success',
+      fetchStatus: 'idle',
+      data: [],
+    })
+
+    unsubscribe()
+  })
+
   test('should replace on refetch', async () => {
     const key = queryKey()
     const observer = new QueryObserver(queryClient, {