+++ packages/vue-query-devtools/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tanstack/vue-query-devtools
 
+## 6.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`b2bd79d`](https://github.com/TanStack/query/commit/b2bd79d0a6b2707461897c426b0d2275a3318e4b)]:
+  - @tanstack/vue-query@5.91.0
+
 ## 5.91.0
 
 ### Minor Changes